+++ litellm-proxy-extras/litellm_proxy_extras/schema.prisma
@@ -128,7 +128,7 @@ model LiteLLM_TeamTable {
     model_max_budget Json @default("{}")
     router_settings  Json? @default("{}")
     team_member_permissions String[] @default([])
-    unified_access_group_ids String[] @default([])
+    access_group_ids String[] @default([])
     policies         String[] @default([])
     model_id Int? @unique // id for LiteLLM_ModelTable -> stores team-level model aliases
     allow_team_guardrail_config Boolean @default(false) // if true, team admin can configure guardrails for this team
@@ -162,7 +162,7 @@ model LiteLLM_DeletedTeamTable {
     model_max_budget        Json     @default("{}")
     router_settings  Json? @default("{}")
     team_member_permissions String[] @default([])
-    unified_access_group_ids String[] @default([])
+    access_group_ids String[] @default([])
     policies                String[] @default([])
     model_id                Int?     // id for LiteLLM_ModelTable -> stores team-level model aliases
     allow_team_guardrail_config Boolean @default(false)
@@ -295,7 +295,7 @@ model LiteLLM_VerificationToken {
     allowed_cache_controls String[] @default([])
     allowed_routes   String[] @default([])
     policies         String[] @default([])
-    unified_access_group_ids String[] @default([])
+    access_group_ids String[] @default([])
     model_spend      Json @default("{}")
     model_max_budget Json @default("{}")
     budget_id String?
@@ -351,7 +351,7 @@ model LiteLLM_DeletedVerificationToken {
     allowed_cache_controls  String[] @default([])
     allowed_routes          String[] @default([])
     policies                String[] @default([])
-    unified_access_group_ids String[] @default([])
+    access_group_ids String[] @default([])
     model_spend             Json     @default("{}")
     model_max_budget        Json     @default("{}")
     router_settings  Json? @default("{}")
@@ -926,7 +926,7 @@ model LiteLLM_PolicyAttachmentTable {
 }
 
 //Unified Access Groups table for storing unified access groups
-model LiteLLM_UnifiedAccessGroup {
+model LiteLLM_AccessGroupTable {
   access_group_id    String   @id @default(uuid())
   access_group_name  String   @unique
   description        String?
