+++ litellm/proxy/proxy_server.py
@@ -393,8 +393,8 @@ from litellm.proxy.management_endpoints.tag_management_endpoints import (
 from litellm.proxy.management_endpoints.team_callback_endpoints import (
     router as team_callback_router,
 )
-from litellm.proxy.management_endpoints.unified_access_groups_endpoints import (
-    router as unified_access_group_router,
+from litellm.proxy.management_endpoints.access_groups_endpoints import (
+    router as access_group_router,
 )
 from litellm.proxy.management_endpoints.team_endpoints import router as team_router
 from litellm.proxy.management_endpoints.team_endpoints import (
@@ -11835,7 +11835,7 @@ app.include_router(enterprise_router)
 app.include_router(ui_discovery_endpoints_router)
 app.include_router(agent_endpoints_router)
 app.include_router(a2a_router)
-app.include_router(unified_access_group_router)
+app.include_router(access_group_router)
 ########################################################
 # MCP Server
 ########################################################
