+++ litellm/proxy/schema.prisma
@@ -128,7 +128,7 @@ model LiteLLM_TeamTable {
     model_max_budget Json @default("{}")
     router_settings  Json? @default("{}")
     team_member_permissions String[] @default([])
-    unified_access_group_ids String[] @default([])
+    access_group_ids String[] @default([])
     policies         String[] @default([])
     model_id Int? @unique // id for LiteLLM_ModelTable -> stores team-level model aliases
     litellm_organization_table LiteLLM_OrganizationTable?   @relation(fields: [organization_id], references: [organization_id])
@@ -161,7 +161,7 @@ model LiteLLM_DeletedTeamTable {
     model_max_budget        Json     @default("{}")
     router_settings  Json? @default("{}")
     team_member_permissions String[] @default([])
-    unified_access_group_ids String[] @default([])
+    access_group_ids String[] @default([])
     policies                String[] @default([])
     model_id                Int?     // id for LiteLLM_ModelTable -> stores team-level model aliases
     
@@ -293,7 +293,7 @@ model LiteLLM_VerificationToken {
     allowed_cache_controls String[] @default([])
     allowed_routes   String[] @default([])
     policies         String[] @default([])
-    unified_access_group_ids String[] @default([])
+    access_group_ids String[] @default([])
     model_spend      Json @default("{}")
     model_max_budget Json @default("{}")
     budget_id String?
@@ -349,7 +349,7 @@ model LiteLLM_DeletedVerificationToken {
     allowed_cache_controls  String[] @default([])
     allowed_routes          String[] @default([])
     policies                String[] @default([])
-    unified_access_group_ids String[] @default([])
+    access_group_ids String[] @default([])
     model_spend             Json     @default("{}")
     model_max_budget        Json     @default("{}")
     router_settings  Json? @default("{}")
@@ -923,7 +923,7 @@ model LiteLLM_PolicyAttachmentTable {
 }
 
 //Unified Access Groups table for storing unified access groups
-model LiteLLM_UnifiedAccessGroup {
+model LiteLLM_AccessGroupTable {
   access_group_id    String   @id @default(uuid())
   access_group_name  String   @unique
   description        String?
